Got all of the functions working
The only thing that I need to do is get the order correctly sorted as per the document
specs went.

The data structure for storing the roster of students must use an interlaced list where one lace is the last name, if same last then go to first, in alphabetic order; the second lace is the grade earned in descending order (highest to lowest), if same grade then go to name; the third lace is the classification, if same classification then go to name.